# benchmarks: use subtests