This avoids duplicating some setup/shutdown work while the go test tool figures out the ideal value for N.

Proof:

```go
func BenchmarkSetupTime(b *testing.B) {
	time.Sleep(5 * time.Second)
	b.Run("benchmark checks", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			time.Sleep(1 * time.Millisecond)
		}
	})
}
func BenchmarkSetupTime2(b *testing.B) {
	time.Sleep(5 * time.Second)
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		time.Sleep(1 * time.Millisecond)
	}
}
```

The first takes 6s the second takes 16.5s