@mrd0ll4r 2021-04-03
Well that is weird. It makes some sense (even though I don't see how it can skip setup work for mutable data), but I found that it changes the result of some benchmarks quite drastically on my setup:

Before, 1 CPU:
```
BenchmarkParseQuery      1328991               902 ns/op
BenchmarkURLParseQuery   1000000              1069 ns/op

BenchmarkHookOverhead/none-v4             357835              3093 ns/op
BenchmarkHookOverhead/none-v6             387542              3118 ns/op
BenchmarkHookOverhead/1nop-v4             386272              3105 ns/op
BenchmarkHookOverhead/1nop-v6             390932              3113 ns/op
BenchmarkHookOverhead/2nop-v4             386751              3110 ns/op
BenchmarkHookOverhead/2nop-v6             385231              3097 ns/op
BenchmarkHookOverhead/3nop-v4             388648              3124 ns/op
BenchmarkHookOverhead/3nop-v6             387634              3090 ns/op
```

After:
```
BenchmarkParseQuery/ParseQuery           2360366               505 ns/op
BenchmarkURLParseQuery/URLParseQuery     2094801               573 ns/op

BenchmarkHookOverhead/none-v4/HandleAnnounce              646209              1595 ns/op
BenchmarkHookOverhead/none-v6/HandleAnnounce              779194              1589 ns/op
BenchmarkHookOverhead/1nop-v4/HandleAnnounce              782712              1588 ns/op
BenchmarkHookOverhead/1nop-v6/HandleAnnounce              748050              1602 ns/op
BenchmarkHookOverhead/2nop-v4/HandleAnnounce              757996              1614 ns/op
BenchmarkHookOverhead/2nop-v6/HandleAnnounce              729048              1613 ns/op
BenchmarkHookOverhead/3nop-v4/HandleAnnounce              765700              1606 ns/op
BenchmarkHookOverhead/3nop-v6/HandleAnnounce              732990              1623 ns/op
```

Before, 16 CPUs:
```
BenchmarkParseQuery-16           1299307               926 ns/op
BenchmarkURLParseQuery-16        1000000              1088 ns/op

BenchmarkHookOverhead/none-v4-16                  360765              3111 ns/op
BenchmarkHookOverhead/none-v6-16                  387217              3191 ns/op
BenchmarkHookOverhead/1nop-v4-16                  395570              3122 ns/op
BenchmarkHookOverhead/1nop-v6-16                  371884              3116 ns/op
BenchmarkHookOverhead/2nop-v4-16                  388794              3112 ns/op
BenchmarkHookOverhead/2nop-v6-16                  390445              3098 ns/op
BenchmarkHookOverhead/3nop-v4-16                  394333              3078 ns/op
BenchmarkHookOverhead/3nop-v6-16                  393452              3070 ns/op
```

After:
```
BenchmarkParseQuery/ParseQuery-16                2340372               509 ns/op
BenchmarkURLParseQuery/URLParseQuery-16          2090116               574 ns/op

BenchmarkHookOverhead/none-v4/HandleAnnounce-16                   641862              1574 ns/op
BenchmarkHookOverhead/none-v6/HandleAnnounce-16                   765442              1576 ns/op
BenchmarkHookOverhead/1nop-v4/HandleAnnounce-16                   760575              1581 ns/op
BenchmarkHookOverhead/1nop-v6/HandleAnnounce-16                   755324              1581 ns/op
BenchmarkHookOverhead/2nop-v4/HandleAnnounce-16                   754470              1586 ns/op
BenchmarkHookOverhead/2nop-v6/HandleAnnounce-16                   769190              1581 ns/op
BenchmarkHookOverhead/3nop-v4/HandleAnnounce-16                   737757              1583 ns/op
BenchmarkHookOverhead/3nop-v6/HandleAnnounce-16                   771490              1588 ns/op
```

The storage benchmarks are mostly unaffected, with a tendency to be higher with the patch applied... so weird. Especially for the values that changed: which ones are the real ones? What's going on?